fixed bug with resizing column
